Remove any node-specific logic from core
