+++ core.ts
@@ -1,5 +1,3 @@
-import { IncomingHttpHeaders, IncomingMessage } from "node:http";
-import { URL } from "node:url";
 import {
     ClientConfig,
     AuthenticationResponse,
@@ -79,7 +77,6 @@ import {
     platformEndpoint,
     screenEndpoint,
     keywordsEndpoint,
-    referrerQueryParameters,
 } from "./constants";
 
 export abstract class Core {
@@ -168,28 +165,6 @@ export abstract class Core {
         }
 
         return await this.performPost(sessionEndpoint, hit);
-    }
-
-    /**
-     * hitFromRequest returns the required data to send a hit to Pirsch for a Node request object.
-     *
-     * @param request the Node request object from the http package.
-     * @returns Hit object containing all necessary fields.
-     */
-    hitFromRequest(request: IncomingMessage): Hit {
-        const url = new URL(request.url ?? "", `${this.protocol}://${this.hostname}`);
-        return {
-            url: url.toString(),
-            ip: request.socket.remoteAddress ?? "",
-            cf_connecting_ip: this.getHeaderWithDefault(request.headers, "cf-connecting-ip"),
-            x_forwarded_for: this.getHeaderWithDefault(request.headers, "x-forwarded-for"),
-            forwarded: this.getHeaderWithDefault(request.headers, "forwarded"),
-            x_real_ip: this.getHeaderWithDefault(request.headers, "x-real-ip"),
-            dnt: this.getHeaderWithDefault(request.headers, "dnt"),
-            user_agent: this.getHeaderWithDefault(request.headers, "user-agent"),
-            accept_language: this.getHeaderWithDefault(request.headers, "accept-language"),
-            referrer: this.getReferrer(request, url),
-        };
     }
 
     /**
@@ -560,37 +535,6 @@ export abstract class Core {
 
             throw error;
         }
-    }
-
-    private getReferrer(request: IncomingMessage, url: URL): string {
-        const referrer =
-            this.getHeader(request.headers, "referer") ?? this.getHeader(request.headers, "referrer") ?? "";
-
-        if (referrer === "") {
-            for (const parameterName of referrerQueryParameters) {
-                const parameter = url.searchParams.get(parameterName);
-
-                if (parameter && parameter !== "") {
-                    return parameter;
-                }
-            }
-        }
-
-        return referrer;
-    }
-
-    private getHeader(headers: IncomingHttpHeaders, name: string): Optional<string> {
-        const header = headers[name];
-
-        if (Array.isArray(header)) {
-            return header.at(0);
-        }
-
-        return header;
-    }
-
-    private getHeaderWithDefault(headers: IncomingHttpHeaders, name: string): string {
-        return this.getHeader(headers, name) ?? "";
     }
 
     protected abstract post<Response, Data extends object = object>(
